+++ sn_networking/CHANGELOG.md
@@ -6,6 +6,42 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.9](https://github.com/maidsafe/safe_network/compare/sn_networking-v0.9.8...sn_networking-v0.9.9) - 2023-10-27
+
+### Added
+- *(networking)* adjust reverification times
+- *(sn_networking)* deterministic store cost order
+
+## [0.9.8](https://github.com/maidsafe/safe_network/compare/sn_networking-v0.9.7...sn_networking-v0.9.8) - 2023-10-26
+
+### Added
+- replicate Spend/Register with same key but different content
+
+### Fixed
+- throw out SplitRecord error for the later on merge
+- client carry out merge when verify register storage
+
+### Other
+- expand replication range
+
+## [0.9.7](https://github.com/maidsafe/safe_network/compare/sn_networking-v0.9.6...sn_networking-v0.9.7) - 2023-10-26
+
+### Fixed
+- add libp2p identity with rand dep for tests
+
+### Other
+- *(networking)* update libp2p for soon to be deprecated changes
+
+## [0.9.6](https://github.com/maidsafe/safe_network/compare/sn_networking-v0.9.5...sn_networking-v0.9.6) - 2023-10-26
+
+### Fixed
+- typos
+
+## [0.9.5](https://github.com/maidsafe/safe_network/compare/sn_networking-v0.9.4...sn_networking-v0.9.5) - 2023-10-26
+
+### Other
+- pass RecordKey by reference
+
 ## [0.9.4](https://github.com/maidsafe/safe_network/compare/sn_networking-v0.9.3...sn_networking-v0.9.4) - 2023-10-24
 
 ### Other