+++ sn_node/CHANGELOG.md
@@ -1865,6 +1865,40 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - *(cli)* chunk files in parallel
 
+## [0.96.7](https://github.com/maidsafe/safe_network/compare/sn_node-v0.96.6...sn_node-v0.96.7) - 2023-10-26
+
+### Other
+- pass RecordKey by reference
+
+## [0.96.8](https://github.com/maidsafe/safe_network/compare/sn_node-v0.96.7...sn_node-v0.96.8) - 2023-10-26
+
+### Fixed
+- typos
+
+## [0.96.9](https://github.com/maidsafe/safe_network/compare/sn_node-v0.96.8...sn_node-v0.96.9) - 2023-10-26
+
+### Fixed
+- add libp2p identity with rand dep for tests
+
+## [0.96.10](https://github.com/maidsafe/safe_network/compare/sn_node-v0.96.9...sn_node-v0.96.10) - 2023-10-26
+
+### Other
+- update dependencies
+
+## [0.96.11](https://github.com/maidsafe/safe_network/compare/sn_node-v0.96.10...sn_node-v0.96.11) - 2023-10-26
+
+### Added
+- replicate Spend/Register with same key but different content
+
+### Other
+- Revert "ci: Register test negation"
+- expand replication range
+
+## [0.96.12](https://github.com/maidsafe/safe_network/compare/sn_node-v0.96.11...sn_node-v0.96.12) - 2023-10-27
+
+### Other
+- update dependencies
+
 ## v0.1.0 (2023-05-04)
 
 ### Chore